Update neonHedgingBasketDraft.sol
Draft notes in the code flow to highlight where and how to add support for basket support. Basket support is the ability to add different ERC20 tokens in the underlying asset basket of a deal. Its mostly applicable and reserved for Lending and not Hedging. Exploring use in hedging ended with a oversized settlement function that also increases costs of settlement.